Code clean up.
Correcting indentation, adding braces, adding serial version ID if
missing, removing unnecessary cast.

+++ src/test/java6/net/finmath/montecarlo/assetderivativevaluation/BachelierModelMonteCarloValuationTest.java
@@ -96,7 +96,7 @@ public class BachelierModelMonteCarloValuationTest {
 
 		long end = System.currentTimeMillis();
 
-		System.out.println("\nCalculation time required: " + ((double)(end-start))/1000.0 + " seconds.");
+		System.out.println("\nCalculation time required: " + (end-start)/1000.0 + " seconds.");
 	}
 
 	public BachelierModelMonteCarloValuationTest() {
@@ -245,7 +245,7 @@ public class BachelierModelMonteCarloValuationTest {
 		for(int i=0; i<100;i++) {
 			System.out.println(i + "\t" + stockAtTimeOne.get(i));
 		}
-		
+
 		double spot = stockAtTimeOne.div(model.getNumeraire(1.0)).mult(model.getNumeraire(model.getTime(0))).getAverage();
 		System.out.println("Expectation of S(1)/N(1)*N(0) = " + spot + " (expected " + initialValue + ")");
 		Assert.assertEquals(initialValue, spot, 2E-3);
@@ -490,7 +490,7 @@ public class BachelierModelMonteCarloValuationTest {
 							)/(2*shift);
 
 			// Calculate the analytic value
-//			double vegaAnalytic	= net.finmath.functions.AnalyticFormulas.blackScholesOptionVega(initialValue, riskFreeRate, volatility, optionMaturity, optionStrike);
+			//			double vegaAnalytic	= net.finmath.functions.AnalyticFormulas.blackScholesOptionVega(initialValue, riskFreeRate, volatility, optionMaturity, optionStrike);
 
 			// Print result
 			System.out.println(numberFormatStrike.format(optionStrike) +
